+++ src/main/java/net/conczin/gui/BookUISupplier.java
@@ -5,10 +5,12 @@ import com.hypixel.hytale.component.ComponentAccessor;
 import com.hypixel.hytale.component.Ref;
 import com.hypixel.hytale.server.core.entity.InteractionContext;
 import com.hypixel.hytale.server.core.entity.entities.player.pages.CustomUIPage;
+import com.hypixel.hytale.server.core.inventory.ItemStack;
 import com.hypixel.hytale.server.core.modules.interaction.interaction.config.server.OpenCustomUIInteraction;
 import com.hypixel.hytale.server.core.universe.PlayerRef;
 import com.hypixel.hytale.server.core.universe.world.storage.EntityStore;
 import net.conczin.utils.RecordCodec;
+import net.conczin.utils.Utils;
 
 import javax.annotation.Nonnull;
 
@@ -25,6 +27,8 @@ public record BookUISupplier(
     @Nonnull
     @Override
     public CustomUIPage tryCreate(Ref<EntityStore> ref, ComponentAccessor<EntityStore> componentAccessor, @Nonnull PlayerRef playerRef, InteractionContext context) {
-        return new BooksGui(playerRef, context.getTargetBlock(), style, background);
+        ItemStack heldItem = context.getHeldItem();
+        boolean isItem = heldItem != null && Utils.getBookSupplier(heldItem) != null;
+        return new BooksGui(playerRef, isItem ? null : context.getTargetBlock(), style, background);
     }
 }